requirements.txt
torch
transformers
accelerate
diffusers
Pillow
gradio
sentencepiece
openai
faiss-cpu
Numpy
